Added getTask, getRepositoriesWithIssues query helpers
